Add input json files via upload
Json files with the input values for the different runs performed. The name of the file indicates: (1) mode in which MIGNON is running (e.g. salmon_star); (2) columns with samples IDs from MignonPatientSelection2.xlsx file; (3) cell type (either CD2 or CD14); (4) samples included in the input file (either all or those that failed).